+++ projects/pastanaga-angular/src/public-api.ts
@@ -63,6 +63,8 @@ export * from './lib/tabs/tabs-list.component';
 export * from './lib/tabs/tab-item.component';
 
 export * from './lib/toast/toast.module';
+export * from './lib/toast/toast.service';
+export * from './lib/toast/toast.model';
 
 export * from './lib/translate/translate.module';
 export * from './lib/translate/translate.pipe';