+++ projects/pastanaga-angular/src/lib/table/table.module.ts
@@ -1,21 +1,30 @@
 import { CommonModule } from '@angular/common';
 import { NgModule } from '@angular/core';
 
-import { TableCellComponent } from './table-cell.component';
+import { TableCellComponent } from './table-cell/table-cell.component';
 import { TableComponent } from './table.component';
 import { TableHeaderDirective } from './table.directives';
-import { TableRowComponent } from './table-row.component';
-import { TableRowHeaderComponent } from './table-row-header.component';
+import { TableRowComponent } from './table-row/table-row.component';
+import { TableRowHeaderComponent } from './table-row-header/table-row-header.component';
+import { TableCellMenuComponent } from './table-cell-menu/table-cell-menu.component';
 
 @NgModule({
     imports: [CommonModule],
-    exports: [TableComponent, TableRowComponent, TableCellComponent, TableRowHeaderComponent, TableHeaderDirective],
+    exports: [
+        TableComponent,
+        TableRowComponent,
+        TableCellComponent,
+        TableRowHeaderComponent,
+        TableHeaderDirective,
+        TableCellMenuComponent,
+    ],
     declarations: [
         TableComponent,
         TableRowComponent,
         TableCellComponent,
         TableRowHeaderComponent,
         TableHeaderDirective,
+        TableCellMenuComponent,
     ],
     providers: [],
 })